# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cider NetEase Lyrics Gateway
  version: 0.2.0
  description: Minimal anonymous search and line-lyric gateway. No login, playback URL, account, Cookie, or unblock routes exist.
  summary: Send Cache-Control no-cache on v1 requests to bypass the short-lived upstream result cache.
servers:
  - url: http://127.0.0.1:3100
paths:
...
  /v1/lyrics/{neteaseId}:
    get:
      operationId: getLyrics
      parameters:
        - name: neteaseId
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9]{1,20}$" }
        - name: durationMs
          in: query
          required: false
          schema: { type: number, exclusiveMinimum: 0 }
      responses:
        "200":
          description: Parsed provider-compatible line lyrics.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LyricsResponse" }
        "400": { $ref: "#/components/responses/InvalidRequest" }
        "404":
          description: Song has no usable line lyrics.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "502": { $ref: "#/components/responses/UpstreamError" }
        "503": { $ref: "#/components/responses/UpstreamError" }
        "504": { $ref: "#/components/responses/Timeout" }
components:
  responses:
    InvalidRequest:
      description: Invalid JSON, schema, ID, or oversized request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    RateLimited:
      description: Per-IP or per-song limit exceeded.
      headers:
        Retry-After: { schema: { type: integer, minimum: 1 } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    UpstreamError:
      description: NetEase is unavailable or rejected the request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Timeout:
      description: Total or upstream deadline elapsed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
  schemas:
...
    LyricLine:
      type: object
      required: [start, end, text, words, empty]
      properties:
        start: { type: number, minimum: 0 }
        end: { type: number, minimum: 0 }
        text: { type: string }
        translation: { type: string }
        words: { type: array, items: { $ref: "#/components/schemas/LyricLine" } }
        empty: { type: boolean }
        isDuet: { type: boolean }
        index: { type: integer, minimum: 0 }
        isCredit: { type: boolean }
    LyricResult:
      type: object
      required: [type, lyrics, source, neteaseId]
      properties:
        type: { enum: [Line, Word, None] }
        lyrics: { type: array, items: { $ref: "#/components/schemas/LyricLine" } }
        source: { const: NetEase }
        neteaseId: { type: string }
    LyricDiagnostics:
      type: object
      required: [fields, unmatchedTranslations, translationCjkRatio, translationIsChinese, parsedRomanizationLines]
      properties:
        fields:
          type: object
          required: [lrc, tlyric, romalrc, yrc]
          properties:
            lrc: { type: boolean }
            tlyric: { type: boolean }
            romalrc: { type: boolean }
            yrc: { type: boolean }
        unmatchedTranslations:
          type: array
          items:
            type: object
            required: [start, text]
            properties:
              start: { type: number }
              text: { type: string }
        translationCjkRatio: { type: number, minimum: 0, maximum: 1 }
        translationIsChinese: { type: boolean }
        parsedRomanizationLines: { type: integer, minimum: 0 }
    LyricsResponse:
      type: object
      required: [schemaVersion, lyrics, diagnostics]
      properties:
        schemaVersion: { const: 1 }
        lyrics: { $ref: "#/components/schemas/LyricResult" }
        diagnostics: { $ref: "#/components/schemas/LyricDiagnostics" }
...
    ErrorResponse:
      type: object
      required: [schemaVersion, error]
      properties:
        schemaVersion: { const: 1 }
        error:
          type: object
          required: [code, message]
          properties:
            code:
              enum: [INVALID_REQUEST, NO_MATCH, NO_LYRICS, TIMEOUT, RATE_LIMITED, UPSTREAM_UNAVAILABLE, SERVICE_ERROR]
            message: { type: string }
            requestId: { type: string }
            retryAfterSeconds: { type: integer, minimum: 1 }